+++ ParallelHashJoins/Part.cs
@@ -9,24 +9,24 @@ namespace ParallelHashJoins
     [Serializable]
     class Part
     {
-        public int pPartKey { get; set; }
+        public Int64 pPartKey { get; set; }
         public string pName { get; set; }
         public string pMFGR { get; set; }
         public string pCategory { get; set; }
         public string pBrand { get; set; }
         public string pColor { get; set; }
         public string pType { get; set; }
-        public int pSize { get; set; }
+        public Int64 pSize { get; set; }
         public string pContainer { get; set; }
 
-        public Part(int pPartKey, 
+        public Part(Int64 pPartKey, 
             string pName, 
             string pMFGR, 
             string pCategory, 
             string pBrand, 
             string pColor, 
             string pType, 
-            int pSize, 
+            Int64 pSize, 
             string pContainer)
         {
             this.pPartKey = pPartKey;
